# Improve performance issue for DELETE command