@seuros 2015-10-29
I wonder why travis didn't pickup the PR.

ping @mceachen