@zzet 2015-10-29
@seuros I updated commit:

```
explain analyze SELECT DISTINCT descendant_id
  FROM (SELECT descendant_id
    FROM "log_hierarchies"
    WHERE ancestor_id = 1022510 or descendant_id = 1022510
  ) AS x;
                                                                QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------
 HashAggregate  (cost=53382.86..53382.88 rows=2 width=4) (actual time=1.008..1.011 rows=22 loops=1)
   Group Key: log_hierarchies.descendant_id
   ->  Bitmap Heap Scan on log_hierarchies  (cost=354.58..53347.42 rows=14179 width=4) (actual time=0.218..0.661 rows=1708 loops=1)
         Recheck Cond: ((ancestor_id = 1022510) OR (descendant_id = 1022510))
         Heap Blocks: exact=65
         ->  BitmapOr  (cost=354.58..354.58 rows=14179 width=0) (actual time=0.202..0.202 rows=0 loops=1)
               ->  Bitmap Index Scan on log_anc_desc_idx  (cost=0.00..86.72 rows=2420 width=0) (actual time=0.045..0.045 rows=56 loops=1)
                     Index Cond: (ancestor_id = 1022510)
               ->  Bitmap Index Scan on log_desc_idx  (cost=0.00..260.76 rows=11759 width=0) (actual time=0.152..0.152 rows=1687 loops=1)
                     Index Cond: (descendant_id = 1022510)
 Planning time: 0.181 ms
 Execution time: 1.067 ms
(12 rows)
```