@zzet 2015-10-29
@mceachen some problems with travis? https://travis-ci.org/mceachen/closure_tree/jobs/88097080#L181