@mceachen 2015-11-02
Thanks so much for the PR! Love the detail, especially the benchmark.
@martin-schmidt fixed the build which I just merged with master, and I'll cut 6.0.0 as soon as this PR builds green on master.